Do not try to destroy runner vm if it's already deleted
Vm can be deleted via UI or respirate might run label again. So if vm is
already deleted, we should not try to delete it. "vm" returns nil, and
fails with "NoMethodError".

+++ prog/vm/github_runner.rb
@@ -184,9 +184,11 @@ class Prog::Vm::GithubRunner < Prog::Base
     rescue Octokit::NotFound
     end
 
-    vm.private_subnets.each { _1.incr_destroy }
-    vm.sshable.destroy
-    vm.incr_destroy
+    if vm
+      vm.private_subnets.each { _1.incr_destroy }
+      vm.sshable.destroy
+      vm.incr_destroy
+    end
 
     hop_wait_vm_destroy
   end
